combine fields and create files, rename ctors to constructors, rename Properties files to submission and respondONly
