+++ VRDR/DeathRecord_responseOnlyProperties.cs
@@ -13,6 +13,9 @@ using Hl7.Fhir.Serialization;
 using Hl7.FhirPath;
 using Newtonsoft.Json;
 
+// DeathRecord_responseOnlyProperties.cs
+//    These fields are used ONLY in coded messages sent from NCHS to EDRS corresponding to TRX and MRE content.
+
 namespace VRDR
 {
     /// <summary>Class <c>DeathRecord</c> models a FHIR Vital Records Death Reporting (VRDR) Death
